New 4.0.0 Release targeting ES6+
The 4+ Releases of Monstache will target ES6+.  This allows us to clean
up things that were working but deprecated or not suggested beyond ES6.
For example, instead of using the MongoDB collection as the ES type
name, the type will be set to _doc on ES 6.2+.  Also, some of the field
names in bulk index requests like _version and _version_type were
renamed to version and version_type.  This allows us to remove the
warnings on those field names.