Make the index-pattern configurable for stateless delete
When using a stateless delete strategy, monstache may perform a search
for the document in Elasticsearch if it was re-routed from the defaults.
This commit makes the scope of that search configurable.  You may use
any Elasticsearch index pattern for config option delete-index-pattern
to do so.  For example, if monstache is configured to only index into
indices a, b, and c you can set this to "a,b,c".  If monstache only
indexes into indices starting with foo you can use "foo*".