Allow global filter functions
This change enables the ability to configure a filter function without a
namespace. In this case the filter will be run on all collections.  The
filter function is now passed the document as the first argument and the
document namespace as the second argument.  You can have a global filter
and a namespace specific filter.  The global filter will be run first
and both filter functions must return true or the document will be
ignored.