Remove the /config endpoint for better security
The http server has always been disabled by default but when enabled
exposes a /config endpoint.  Since this endpoint may be misconfigured by
accident and leak connection information, this commit removes it.  A
secure version may be added back in the future if a use case for this
endpoint presents itself.  The -print-config option is still available
to view JSON formatted settings.