Fix for issue #89
This commit makes the oplog timestamp and date field names configurable
and also makes the oplog date field format configurable.  The previous
default values for the field names were _oplog_ts and _oplog_date.
Because the defaults made it impossible to reference the fields in Kibana
this commit also changes the default field names to oplog_ts and
oplog_date.  To restore the previous defaults or use different field
names, use the configuration options oplog-ts-field-name and
oplog-date-field-name. To use a different format for the date field use
the configuration option oplog-date-field-format.  The format must
adhere to the go standard library date format.  Closes issue #89.