Post processor and replacement features
Run Process methods of golang plugins off the main loop for performance.

New replacement config option to do simple replacement of one doc with
another before indexing. This is useful if your direct-read-namespaces
point to a view and you also want change events in the underlying
collections to be read from the view before indexing.